requirements.txt
python-telegram-bot==20.6
google-auth
google-auth-oauthlib
google-api-python-client
fastkml
lxml